application/requirements.txt
Flask==2.3.3
requests==2.25.1
PyYAML==5.3.1
cryptography==3.3.2

